+++ extensions/minigraph-playground-engine/src/main/java/com/accenture/minigraph/rest/InspectStateMachine.java
@@ -1,0 +1,53 @@
+/*
+
+    Copyright 2018-2026 Accenture Technology
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
+package com.accenture.minigraph.rest;
+
+import com.accenture.minigraph.services.GraphCommandService;
+import org.platformlambda.core.annotations.OptionalService;
+import org.platformlambda.core.annotations.PreLoad;
+import org.platformlambda.core.exception.AppException;
+import org.platformlambda.core.models.AsyncHttpRequest;
+import org.platformlambda.core.models.EventEnvelope;
+import org.platformlambda.core.models.TypedLambdaFunction;
+
+import java.util.List;
+import java.util.Map;
+
+@OptionalService("app.env=dev")
+@PreLoad(route = "inspect.state.machine", instances = 10)
+public class InspectStateMachine implements TypedLambdaFunction<AsyncHttpRequest, Object> {
+
+    @Override
+    public Object handleEvent(Map<String, String> headers, AsyncHttpRequest input, int instance) {
+        var id = input.getPathParameter("id");
+        var key = input.getPathParameter("key");
+        if (id == null) {
+            throw new IllegalArgumentException("Missing path parameter: id");
+        }
+        if (key == null) {
+            throw new IllegalArgumentException("Missing path parameter: key");
+        }
+        var content = GraphCommandService.downloadContent(id, key);
+        if (content instanceof Map || content instanceof List) {
+            return new EventEnvelope().setHeader("Content-Type", "application/json").setBody(content);
+        } else {
+            throw new AppException(404, "Not found");
+        }
+    }
+}