+++ extensions/minigraph-playground-engine/src/main/java/com/accenture/minigraph/services/GraphCommandService.java
@@ -21,6 +21,7 @@ package com.accenture.minigraph.services;
 import com.accenture.minigraph.common.GraphLambdaFunction;
 import com.accenture.minigraph.models.GraphInstance;
 import com.jayway.jsonpath.InvalidPathException;
+import org.platformlambda.core.annotations.OptionalService;
 import org.platformlambda.core.annotations.PreLoad;
 import org.platformlambda.core.graph.MiniGraph;
 import org.platformlambda.core.models.EventEnvelope;
@@ -45,6 +46,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
+@OptionalService("app.env=dev")
 @PreLoad(route = GraphCommandService.ROUTE, instances=50)
 public class GraphCommandService extends GraphLambdaFunction {
     public static final String ROUTE = "graph.command.service";
@@ -53,6 +55,7 @@ public class GraphCommandService extends GraphLambdaFunction {
     private static final String DEFAULT_DEPLOY_DIR = "classpath:/graph";
     private static final String PLAYGROUND = "playground";
     private static final String INVALID_GRAPH_NAME = "Invalid filename - must be a-z, A-Z, 0-9 with optional hyphen";
+    private static final long MAX_BUFFER_SIZE = 62 * 1024L;
     private final AtomicInteger counter = new AtomicInteger();
     private final File tempDir;
     private final String deployedGraphLocation;
@@ -214,6 +217,18 @@ public class GraphCommandService extends GraphLambdaFunction {
             var po = EventEmitter.getInstance();
             po.send(outRoute, "Mock data loaded into 'input.body' namespace");
             return true;
+        }
+    }
+
+    public static Object downloadContent(String id, String key) {
+        var route = id.replace('-', '.');
+        var inRoute = route + ".in";
+        var instance = graphInstances.get(inRoute);
+        if (instance != null) {
+            var stateMachine = instance.stateMachine;
+            return stateMachine.getElement(key);
+        } else {
+            return null;
         }
     }
 
@@ -444,7 +459,18 @@ public class GraphCommandService extends GraphLambdaFunction {
     private void handleInspectCommand(PostOffice po, String inRoute, String outRoute, String key) {
         var stateMachine = getGraphInstance(inRoute).stateMachine;
         var value = stateMachine.getElement(key, "null");
-        po.send(new EventEnvelope().setTo(outRoute).setBody(Map.of("inspect", key, "outcome", value)));
+        if (value instanceof Map || value instanceof List) {
+            var text = SimpleMapper.getInstance().getMapper().writeValueAsString(value);
+            if (text.length() > MAX_BUFFER_SIZE) {
+                var name = getTempGraphName(inRoute);
+                po.send(new EventEnvelope().setTo(outRoute).setBody("Large payload ("+ text.length()
+                        +") -> GET /api/inspect/"+ name+"/"+key));
+            } else {
+                po.send(new EventEnvelope().setTo(outRoute).setBody(Map.of("inspect", key, "outcome", value)));
+            }
+        } else {
+            po.send(new EventEnvelope().setTo(outRoute).setBody(Map.of("inspect", key, "outcome", value)));
+        }
     }
 
     private void handleExecuteCommand(PostOffice po, String inRoute, String outRoute, List<String> words) {
@@ -778,7 +804,7 @@ public class GraphCommandService extends GraphLambdaFunction {
             throw new IllegalArgumentException("Did you forget to create an end node?");
         }
         var line = lines.get(i);
-        var sep = line.indexOf(MAP_TO);
+        var sep = line.lastIndexOf(MAP_TO);
         if (sep == -1) {
             throw new IllegalArgumentException("Invalid data mapping entry. e.g. 'source -> target'");
         }